+++ Evergreen/Apps/Get-MicrosoftTeams.ps1
@@ -12,7 +12,7 @@ Function Get-MicrosoftTeams {
             https://teams.microsoft.com/desktopclient/installer/windows/arm64
     #>
     [OutputType([System.Management.Automation.PSObject])]
-    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseSingularNouns", "")]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseSingularNouns", "", Justification="Product name is a plural")]
     [CmdletBinding(SupportsShouldProcess = $False)]
     param (
         [Parameter(Mandatory = $False, Position = 0)]