+++ Evergreen/Public/New-EvergreenLibrary.ps1
@@ -60,9 +60,8 @@ function New-EvergreenLibrary {
             Write-Verbose -Message "Library exists: $Path."
         }
         else {
-            $Library.Name = $Name
-
             try {
+                $Library.Name = $Name
                 $Library | ConvertTo-Json -Depth 10 | Out-File -FilePath $LibraryFile -Encoding "Utf8" -NoNewline -Verbose:$VerbosePreference
             }
             catch {